refactor(cloudformation): move the Events types into a per-service provisioner
AWS::Events::Rule, EventBus and EventBusPolicy move out of the legacy monolith
into EventsCfnProvisioner, continuing the decomposition. The block was already
contiguous and self-contained: 18 methods and five constants, none of them
referenced from outside, and EventBridgeService was the monolith's only user, so
the field, its constructor parameter and both imports go too.

Behaviour preserving. The event bus arm in particular already carried real logic
worth keeping intact: rollback ownership, the created-time check that refuses to
delete a bus recreated out of band, managed tag and policy tracking, and the
rejection of an unsupported mutable update. All of it moves unchanged.

Rule and EventBus were listed in DELETE_NEEDS_STACK_RESOURCE because deleting
them needs stored attributes rather than the physical id alone: a rule remembers
the custom bus it lives on, and a bus remembers the creation time proving the
stack still owns it. That logic moves into the provisioner's
delete(StackResource, String) override and both entries leave the set, which is
what CfnDeletePrecedenceTest gates. EventBusPolicy keeps the id-only form, its
composite busName|statementId id being enough.

The four existing Events test classes needed no changes: they build the
provisioner through CfnProvisionerFixture, so wiring the new provisioner there is
all it took for them to exercise it. 40 tests across them still pass, and the
schema-coverage gate finds no new attribute gaps.

Monolith 6,457 to 5,954 lines, legacy types 47 to 44.
