+++ src/main/java/io/github/hectorvent/floci/services/cloudformation/CloudFormationResourceProvisioner.java
@@ -20,9 +20,7 @@ import io.github.hectorvent.floci.services.cloudformation.provisioners.CfnResour
 import io.github.hectorvent.floci.services.cloudformation.provisioners.Ec2SecurityGroupRuleCfnProvisioner;
 import io.github.hectorvent.floci.services.cloudformation.provisioners.UpdateCleanupResult;
 import io.github.hectorvent.floci.services.dynamodb.DynamoDbService;
-import io.github.hectorvent.floci.services.eventbridge.EventBridgeService;
 import io.github.hectorvent.floci.services.eventbridge.model.BatchParameters;
-import io.github.hectorvent.floci.services.eventbridge.model.EventBus;
 import io.github.hectorvent.floci.services.eventbridge.model.InputTransformer;
 import io.github.hectorvent.floci.services.eventbridge.model.RuleState;
 import io.github.hectorvent.floci.services.eventbridge.model.SqsParameters;
@@ -126,13 +124,6 @@ public class CloudFormationResourceProvisioner {
     private static final String INLINE_CLEANUP_GROUP_TARGETS_ATTR = "__FlociInlineCleanupGroupTargets";
     private static final String SFN_NAME_MODE_ATTR = "FlociStepFunctionsNameMode";
     static final String SFN_UPDATE_SNAPSHOT_ATTR = "__FlociStepFunctionsUpdateSnapshot";
-    private static final String EVENT_BUS_CREATED_TIME_ATTR = "FlociEventBusCreatedTime";
-    private static final String EVENT_BUS_MANAGED_TAG_KEYS_ATTR = "FlociEventBusManagedTagKeys";
-    private static final String EVENT_BUS_MANAGED_POLICY_ATTR = "FlociEventBusManagedPolicy";
-    private static final Set<String> EVENT_BUS_SUPPORTED_PROPERTIES =
-            Set.of("Name", "Description", "Tags", "Policy");
-    private static final Pattern EVENT_BUS_TAG_PATTERN =
-            Pattern.compile("[\\p{L}\\p{N}\\p{Z}_.:/=+\\-@]*");
     private static final String NAME_MODE_EXPLICIT = "explicit";
     private static final String NAME_MODE_GENERATED = "generated";
     private static final int GENERATED_NAME_SUFFIX_LENGTH = 12;
@@ -177,8 +168,6 @@ public class CloudFormationResourceProvisioner {
             "AWS::ApiGatewayV2::Authorizer",
             "AWS::CloudFormation::CustomResource",
             "AWS::EKS::Nodegroup",
-            "AWS::Events::EventBus",
-            "AWS::Events::Rule",
             "AWS::IAM::ManagedPolicy",
             "AWS::IAM::Policy",
             "AWS::SecretsManager::SecretTargetAttachment",
@@ -210,9 +199,6 @@ public class CloudFormationResourceProvisioner {
             "AWS::EC2::SecurityGroup",
             "AWS::EKS::Cluster",
             "AWS::EKS::Nodegroup",
-            "AWS::Events::EventBus",
-            "AWS::Events::EventBusPolicy",
-            "AWS::Events::Rule",
             "AWS::IAM::AccessKey",
             "AWS::IAM::InstanceProfile",
             "AWS::IAM::ManagedPolicy",
@@ -247,7 +233,6 @@ public class CloudFormationResourceProvisioner {
     private final LambdaService lambdaService;
     private final IamService iamService;
     private final SecretsManagerService secretsManagerService;
-    private final EventBridgeService eventBridgeService;
     private final ApiGatewayService apiGatewayService;
     private final ApiGatewayV2Service apiGatewayV2Service;
     private final LambdaLayerService lambdaLayerService;
@@ -275,7 +260,6 @@ public class CloudFormationResourceProvisioner {
                                              LambdaService lambdaService, IamService iamService,
                                              SsmService ssmService, KmsService kmsService,
                                              SecretsManagerService secretsManagerService,
-                                             EventBridgeService eventBridgeService,
                                              ApiGatewayService apiGatewayService,
                                              ApiGatewayV2Service apiGatewayV2Service,
                                              EcrService ecrService,
@@ -305,7 +289,6 @@ public class CloudFormationResourceProvisioner {
         this.lambdaService = lambdaService;
         this.iamService = iamService;
         this.secretsManagerService = secretsManagerService;
-        this.eventBridgeService = eventBridgeService;
         this.apiGatewayService = apiGatewayService;
         this.apiGatewayV2Service = apiGatewayV2Service;
         this.lambdaLayerService = lambdaLayerService;
@@ -379,9 +362,6 @@ public class CloudFormationResourceProvisioner {
                 case "AWS::SecretsManager::SecretTargetAttachment" ->
                         provisionSecretTargetAttachment(resource, properties, engine, region, stackName);
                 case "AWS::Route53::RecordSet" -> provisionRoute53RecordSet(resource, properties, engine);
-                case "AWS::Events::Rule" -> provisionEventBridgeRule(resource, properties, engine, region, stackName);
-                case "AWS::Events::EventBus" -> provisionEventBridgeEventBus(resource, properties, engine, region);
-                case "AWS::Events::EventBusPolicy" -> provisionEventBusPolicy(resource, properties, engine, region);
                 case "AWS::ApiGateway::RestApi" -> provisionApiGatewayRestApi(resource, properties, engine, region, accountId, stackName);
                 case "AWS::ApiGateway::Resource" -> provisionApiGatewayResource(resource, properties, engine, region);
                 case "AWS::ApiGateway::Authorizer" -> provisionApiGatewayAuthorizer(resource, properties, engine, region);
@@ -600,14 +580,6 @@ public class CloudFormationResourceProvisioner {
             }
             return;
         }
-        // Rule deletion needs the rule's event bus (stored as an attribute at provision time),
-        // which the type/physicalId delete path can't provide; a custom-bus rule looked up under
-        // the default bus would silently no-op and leave the rule (and its bus) live.
-        if ("AWS::Events::Rule".equals(resourceType)) {
-            deleteEventBridgeRuleSafe(resource.getPhysicalId(),
-                    resource.getAttributes().get("EventBusName"), region);
-            return;
-        }
         // Authorizer deletion needs the api id (a stored attribute, not the physical id, which is
         // the authorizer id) — same shape as the Nodegroup case above. Without this, the generic
         // type/physicalId delete path has no case for this type at all and silently no-ops,
@@ -633,10 +605,6 @@ public class CloudFormationResourceProvisioner {
         // policy before IAM's DeletePolicy operation. The type/physicalId path lacks that state.
         if ("AWS::IAM::ManagedPolicy".equals(resourceType)) {
             deleteManagedPolicy(resource);
-            return;
-        }
-        if ("AWS::Events::EventBus".equals(resourceType)) {
-            deleteEventBusSafe(resource, region);
             return;
         }
         throw new IllegalStateException("DELETE_NEEDS_STACK_RESOURCE lists " + resourceType
@@ -673,9 +641,6 @@ public class CloudFormationResourceProvisioner {
                     "SecretTargetAttachment deletion requires the StackResource metadata that records its managed fields.",
                     400);
             // No bus context on the type/physicalId path (e.g. CREATE-rollback); targets the default bus.
-            case "AWS::Events::Rule" -> deleteEventBridgeRuleSafe(physicalId, null, region);
-            case "AWS::Events::EventBus" -> deleteEventBusSafe(physicalId, region);
-            case "AWS::Events::EventBusPolicy" -> removeEventBusPolicySafe(physicalId, region);
             case "AWS::ApiGateway::RestApi" -> apiGatewayService.deleteRestApi(region, physicalId);
             case "AWS::ApiGatewayV2::Api" -> apiGatewayV2Service.deleteApi(region, physicalId);
             case "AWS::StepFunctions::StateMachine" -> stepFunctionsService.deleteStateMachine(physicalId);
@@ -3306,474 +3271,6 @@ public class CloudFormationResourceProvisioner {
         return password;
     }
 
-    // ── EventBridge ─────────────────────────────────────────────────────────
-
-    private void provisionEventBridgeRule(StackResource r, JsonNode props, CloudFormationTemplateEngine engine,
-                                          String region, String stackName) {
-        String ruleName = resolveOptional(props, "Name", engine);
-        if (ruleName == null || ruleName.isBlank()) {
-            ruleName = generatePhysicalName(stackName, r.getLogicalId(), 64, false);
-        }
-
-        String busName = resolveOptional(props, "EventBusName", engine);
-        String description = resolveOptional(props, "Description", engine);
-        String roleArn = resolveOptional(props, "RoleArn", engine);
-        String scheduleExpression = resolveOptional(props, "ScheduleExpression", engine);
-
-        String eventPattern = null;
-        if (props != null && props.has("EventPattern") && !props.get("EventPattern").isNull()) {
-            JsonNode patternNode = engine.resolveNode(props.get("EventPattern"));
-            eventPattern = patternNode.toString();
-        }
-
-        String stateStr = resolveOptional(props, "State", engine);
-        RuleState state = "DISABLED".equals(stateStr) ? RuleState.DISABLED : RuleState.ENABLED;
-
-        var rule = eventBridgeService.putRule(ruleName, busName, eventPattern, scheduleExpression,
-                state, description, roleArn, Map.of(), region);
-        r.setPhysicalId(ruleName);
-        r.getAttributes().put("Arn", rule.getArn());
-        // A rule on a custom bus is keyed by that bus; remember it so the resource delete can target
-        // the right bus (the physical id is only the rule name, which resolves to the default bus).
-        if (busName != null && !busName.isBlank()) {
-            r.getAttributes().put("EventBusName", busName);
-        }
-
-        // Provision inline targets
-        if (props != null && props.has("Targets")) {
-            List<Target> targets = new ArrayList<>();
-            for (JsonNode targetNode : props.get("Targets")) {
-                JsonNode resolved = engine.resolveNode(targetNode);
-                String targetId = resolved.path("Id").asText(null);
-                String targetArn = resolved.path("Arn").asText(null);
-                String input = resolved.path("Input").asText(null);
-                String inputPath = resolved.path("InputPath").asText(null);
-                if (targetId != null && targetArn != null) {
-                    Target target = new Target(targetId, targetArn, input, inputPath);
-                    target.setInputTransformer(InputTransformer.fromJson(resolved.path("InputTransformer")));
-                    JsonNode sqsParamsNode = resolved.path("SqsParameters");
-                    if (!sqsParamsNode.isMissingNode() && sqsParamsNode.isObject()) {
-                        String messageGroupId = sqsParamsNode.path("MessageGroupId").asText(null);
-                        if (messageGroupId != null) {
-                            SqsParameters sqsParameters = new SqsParameters();
-                            sqsParameters.setMessageGroupId(messageGroupId);
-                            target.setSqsParameters(sqsParameters);
-                        }
-                    }
-                    JsonNode batchParamsNode = resolved.path("BatchParameters");
-                    if (!batchParamsNode.isMissingNode() && batchParamsNode.isObject()) {
-                        JsonNode arrayProperties = batchParamsNode.path("ArrayProperties");
-                        BatchParameters batchParameters = new BatchParameters();
-                        batchParameters.setJobDefinition(batchParamsNode.path("JobDefinition").asText(null));
-                        batchParameters.setJobName(batchParamsNode.path("JobName").asText(null));
-                        if (arrayProperties.isObject()) {
-                            batchParameters.setArrayProperties(jsonObjectToMap(arrayProperties));
-                        }
-                        if (batchParamsNode.has("RetryStrategy")) {
-                            batchParameters.setRetryStrategy(batchParamsNode.get("RetryStrategy"));
-                        }
-                        target.setBatchParameters(batchParameters);
-                    }
-                    targets.add(target);
-                }
-            }
-            if (!targets.isEmpty()) {
-                eventBridgeService.putTargets(ruleName, busName, targets, region);
-            }
-        }
-    }
-
-    /**
-     * Provisions an {@code AWS::Events::EventBus} (a custom EventBridge event bus). Without this the
-     * resource would fall through to the generic stub, which assigns a physical id but never registers
-     * the bus with the EventBridge service — so any {@code AWS::Events::Rule} (or PutEvents) targeting
-     * the bus fails "EventBus not found". Per the AWS spec, {@code Ref} returns the bus <em>name</em>
-     * (not the ARN), so the physical id is the name; {@code Fn::GetAtt "Arn"} exposes the ARN.
-     */
-    private void provisionEventBridgeEventBus(StackResource r, JsonNode props,
-                                              CloudFormationTemplateEngine engine, String region) {
-        validateEventBusProperties(props);
-        String existingBusName = r.getPhysicalId();
-        String busName = resolveOptional(props, "Name", engine);
-        validateEventBusName(busName);
-        if (existingBusName != null && !existingBusName.equals(busName)) {
-            throw new AwsException("ValidationError",
-                    "Updating EventBus Name requires resource replacement, which is not supported.", 400);
-        }
-        String description = resolveOptional(props, "Description", engine);
-        if (description != null && description.length() > 512) {
-            throw new AwsException("ValidationError",
-                    "AWS::Events::EventBus Description must not exceed 512 characters.", 400);
-        }
-        Map<String, String> tags = parseEventBusTags(
-                props != null ? props.get("Tags") : null, engine);
-        JsonNode policy = resolveEventBusPolicy(props, engine);
-
-        EventBus bus;
-        boolean createdBus = false;
-        try {
-            bus = eventBridgeService.createEventBus(busName, description, tags, region);
-            createdBus = true;
-            r.getAttributes().put(CfnRollback.ROLLBACK_OWNED_ATTR, "true");
-        } catch (AwsException e) {
-            boolean stackAlreadyOwnsBus = existingBusName != null && existingBusName.equals(busName);
-            if (!stackAlreadyOwnsBus || !"ResourceAlreadyExistsException".equals(e.getErrorCode())) {
-                throw e;
-            }
-            bus = eventBridgeService.describeEventBus(busName, region);
-            // A missing created-time means ownership was never tracked, not that it changed: stacks
-            // provisioned before this attribute existed are restored without it. Refusing there would
-            // wedge every later UpdateStack, including no-op ones. Only a recorded time that actually
-            // disagrees means the bus was recreated out of band and belongs to its new owner.
-            String existingCreatedTime = r.getAttributes().get(EVENT_BUS_CREATED_TIME_ATTR);
-            String actualCreatedTime = eventBusCreatedTime(bus);
-            if (existingCreatedTime != null && !existingCreatedTime.equals(actualCreatedTime)) {
-                r.getAttributes().remove(CfnRollback.ROLLBACK_OWNED_ATTR);
-                throw e;
-            }
-            validateEventBusMutablePropertiesUnchanged(r, bus, description, tags, policy);
-        }
-
-        // Record identity before applying the policy: rollbackCreatedResources skips any resource
-        // whose physicalId is still null, so a putPermission failure after the bus exists would
-        // otherwise orphan it with no way for rollback to find it.
-        r.setPhysicalId(busName);              // Ref → EventBus name (AWS-faithful)
-        r.getAttributes().put("Arn", bus.getArn());
-        r.getAttributes().put("Name", busName);
-        r.getAttributes().put(EVENT_BUS_CREATED_TIME_ATTR, eventBusCreatedTime(bus));
-        recordEventBusManagedTagKeys(r, tags.keySet());
-        recordEventBusManagedPolicy(r, policy);
-
-        // Apply an optional inline resource policy only during creation. Updating it is rejected by
-        // validateEventBusMutablePropertiesUnchanged until stack updates can roll back live resource
-        // mutations transactionally.
-        if (createdBus && !policy.isNull()) {
-            eventBridgeService.putPermission(busName, null, null, null, null, policy.toString(), region);
-        }
-    }
-
-    private void validateEventBusProperties(JsonNode props) {
-        if (props == null || props.isNull()) {
-            return;
-        }
-        if (!props.isObject()) {
-            throw new AwsException("ValidationError",
-                    "AWS::Events::EventBus Properties must be an object.", 400);
-        }
-        List<String> unsupported = new ArrayList<>();
-        props.fieldNames().forEachRemaining(name -> {
-            if (!EVENT_BUS_SUPPORTED_PROPERTIES.contains(name)) {
-                unsupported.add(name);
-            }
-        });
-        if (!unsupported.isEmpty()) {
-            Collections.sort(unsupported);
-            throw new AwsException("ValidationError",
-                    "Unsupported AWS::Events::EventBus properties: "
-                            + String.join(", ", unsupported), 400);
-        }
-    }
-
-    private void validateEventBusName(String busName) {
-        if (busName == null || busName.isBlank()) {
-            throw new AwsException("ValidationError",
-                    "Name is required for AWS::Events::EventBus.", 400);
-        }
-        if (busName.length() > 256
-                || !busName.matches("[.\\-_A-Za-z0-9]+")
-                || "default".equals(busName)) {
-            throw new AwsException("ValidationError",
-                    "Invalid custom event bus Name: " + busName, 400);
-        }
-    }
-
-    private Map<String, String> parseEventBusTags(
-            JsonNode tagsNode, CloudFormationTemplateEngine engine) {
-        if (tagsNode == null || tagsNode.isNull()) {
-            return Map.of();
-        }
-        JsonNode resolvedTags = engine.resolveNode(tagsNode);
-        if (!resolvedTags.isArray()) {
-            throw new AwsException("ValidationError",
-                    "AWS::Events::EventBus Tags must be an array.", 400);
-        }
-        if (resolvedTags.size() > 50) {
-            throw new AwsException("ValidationError",
-                    "AWS::Events::EventBus supports at most 50 tags.", 400);
-        }
-        Map<String, String> tags = new LinkedHashMap<>();
-        for (JsonNode entry : resolvedTags) {
-            if (!entry.isObject()) {
-                throw new AwsException("ValidationError",
-                        "Each AWS::Events::EventBus tag must be an object.", 400);
-            }
-            String key = entry.path("Key").asText(null);
-            String value = entry.path("Value").asText(null);
-            if (key == null || key.isEmpty() || key.length() > 128) {
-                throw new AwsException("ValidationError",
-                        "Event bus tag Key must contain 1 to 128 characters.", 400);
-            }
-            if (key.regionMatches(true, 0, "aws:", 0, 4)) {
-                throw new AwsException("ValidationError",
-                        "Event bus tag Key must not use the reserved aws: prefix.", 400);
-            }
-            if (!EVENT_BUS_TAG_PATTERN.matcher(key).matches()) {
-                throw new AwsException("ValidationError",
-                        "Event bus tag Key contains unsupported characters.", 400);
-            }
-            if (value == null || value.length() > 256) {
-                throw new AwsException("ValidationError",
-                        "Event bus tag Value must contain at most 256 characters.", 400);
-            }
-            if (!EVENT_BUS_TAG_PATTERN.matcher(value).matches()) {
-                throw new AwsException("ValidationError",
-                        "Event bus tag Value contains unsupported characters.", 400);
-            }
-            if (tags.putIfAbsent(key, value) != null) {
-                throw new AwsException("ValidationError",
-                        "Duplicate event bus tag Key: " + key, 400);
-            }
-        }
-        return tags;
-    }
-
-    private void validateEventBusMutablePropertiesUnchanged(
-            StackResource resource, EventBus bus, String requestedDescription,
-            Map<String, String> requestedTags, JsonNode requestedPolicy) {
-        if (!Objects.equals(bus.getDescription(), requestedDescription)) {
-            throw unsupportedEventBusMutableUpdate();
-        }
-
-        // Stacks persisted by the older EventBus provisioner have no managed-key metadata. There
-        // is no reliable way to distinguish their CloudFormation tags from tags added out of band,
-        // so allow this one-time adoption and start tracking the requested keys afterwards.
-        if (resource.getAttributes().containsKey(EVENT_BUS_MANAGED_TAG_KEYS_ATTR)) {
-            Map<String, String> currentManagedTags = new LinkedHashMap<>();
-            for (String key : eventBusManagedTagKeys(resource)) {
-                if (bus.getTags().containsKey(key)) {
-                    currentManagedTags.put(key, bus.getTags().get(key));
-                }
-            }
-            if (!currentManagedTags.equals(requestedTags)) {
-                throw unsupportedEventBusMutableUpdate();
-            }
-        }
-
-        String managedPolicy = resource.getAttributes().get(EVENT_BUS_MANAGED_POLICY_ATTR);
-        JsonNode policyToCompare = managedPolicy != null
-                ? parseEventBusPolicy(managedPolicy, "stored CloudFormation metadata")
-                : parseEventBusPolicy(bus.getPolicy(), "the existing event bus");
-        if (managedPolicy != null || !requestedPolicy.isNull()) {
-            if (!policyToCompare.equals(requestedPolicy)) {
-                throw unsupportedEventBusMutableUpdate();
-            }
-        }
-    }
-
-    private JsonNode resolveEventBusPolicy(JsonNode props, CloudFormationTemplateEngine engine) {
-        if (props == null || !props.has("Policy") || props.get("Policy").isNull()) {
-            return JsonNodeFactory.instance.nullNode();
-        }
-        return engine.resolveNode(props.get("Policy"));
-    }
-
-    private JsonNode parseEventBusPolicy(String policy, String source) {
-        if (policy == null) {
-            return JsonNodeFactory.instance.nullNode();
-        }
-        try {
-            JsonNode parsed = objectMapper.readTree(policy);
-            return parsed != null ? parsed : JsonNodeFactory.instance.nullNode();
-        } catch (Exception e) {
-            throw new AwsException("InternalFailure",
-                    "Invalid EventBus policy in " + source + ": " + e.getMessage(), 500);
-        }
-    }
-
-    private void recordEventBusManagedPolicy(StackResource resource, JsonNode policy) {
-        try {
-            resource.getAttributes().put(EVENT_BUS_MANAGED_POLICY_ATTR,
-                    objectMapper.writeValueAsString(policy));
-        } catch (Exception e) {
-            throw new AwsException("InternalFailure",
-                    "Failed to store EventBus managed-policy metadata: " + e.getMessage(), 500);
-        }
-    }
-
-    private AwsException unsupportedEventBusMutableUpdate() {
-        return new AwsException("ValidationError",
-                "Updating AWS::Events::EventBus Description, Tags, or Policy is not supported "
-                        + "until transactional rollback is available.", 400);
-    }
-
-    private Set<String> eventBusManagedTagKeys(StackResource resource) {
-        String value = resource.getAttributes().get(EVENT_BUS_MANAGED_TAG_KEYS_ATTR);
-        if (value == null || value.isBlank()) {
-            return Set.of();
-        }
-        try {
-            JsonNode keys = objectMapper.readTree(value);
-            if (!keys.isArray()) {
-                throw new IllegalArgumentException("managed tag keys are not an array");
-            }
-            Set<String> result = new HashSet<>();
-            keys.forEach(key -> result.add(key.asText()));
-            return result;
-        } catch (Exception e) {
-            throw new AwsException("InternalFailure",
-                    "Invalid stored EventBus managed-tag metadata: " + e.getMessage(), 500);
-        }
-    }
-
-    private void recordEventBusManagedTagKeys(StackResource resource, Set<String> keys) {
-        try {
-            resource.getAttributes().put(
-                    EVENT_BUS_MANAGED_TAG_KEYS_ATTR,
-                    objectMapper.writeValueAsString(new TreeSet<>(keys)));
-        } catch (Exception e) {
-            throw new AwsException("InternalFailure",
-                    "Failed to store EventBus managed-tag metadata: " + e.getMessage(), 500);
-        }
-    }
-
-    private String eventBusCreatedTime(EventBus bus) {
-        return bus.getCreatedTime() != null ? bus.getCreatedTime().toString() : "";
-    }
-
-    private void deleteEventBridgeRuleSafe(String ruleName, String busName, String region) {
-        try {
-            // Remove all targets before deleting the rule (busName scopes the lookup to the rule's bus).
-            var targets = eventBridgeService.listTargetsByRule(ruleName, busName, region);
-            if (!targets.isEmpty()) {
-                List<String> targetIds = targets.stream().map(Target::getId).toList();
-                eventBridgeService.removeTargets(ruleName, busName, targetIds, region);
-            }
-            eventBridgeService.deleteRule(ruleName, busName, region);
-        } catch (AwsException e) {
-            // An already-deleted rule is the one failure that genuinely means "done". Anything else
-            // is a real error worth surfacing rather than hiding behind a debug line.
-            if (!"ResourceNotFoundException".equals(e.getErrorCode())) {
-                throw e;
-            }
-            LOG.debugv("EventBridge rule already gone, treating as deleted: {0}", ruleName);
-        } catch (Exception e) {
-            throw new AwsException("InternalFailure",
-                    "Could not delete EventBridge rule " + ruleName + ": " + e.getMessage(), 500);
-        }
-    }
-
-    private void deleteEventBusSafe(String busName, String region) {
-        try {
-            eventBridgeService.deleteEventBus(busName, region);
-        } catch (AwsException e) {
-            if (!"ResourceNotFoundException".equals(e.getErrorCode())) {
-                throw e;
-            }
-            LOG.debugv("Event bus already gone, treating as deleted: {0}", busName);
-        }
-    }
-
-    private void deleteEventBusSafe(StackResource resource, String region) {
-        String busName = resource.getPhysicalId();
-        EventBus bus;
-        try {
-            bus = eventBridgeService.describeEventBus(busName, region);
-        } catch (AwsException e) {
-            if ("ResourceNotFoundException".equals(e.getErrorCode())) {
-                LOG.debugv("Event bus already gone, treating as deleted: {0}", busName);
-                return;
-            }
-            throw e;
-        }
-
-        // A missing attribute means ownership was never tracked, not that it changed: stacks
-        // provisioned before this attribute existed are restored from cloudformation-stacks.json
-        // without it. Refusing there would leave every such stack permanently in DELETE_FAILED, so
-        // fall back to the pre-tracking behaviour of deleting what the stack recorded it created.
-        String expectedCreatedTime = resource.getAttributes().get(EVENT_BUS_CREATED_TIME_ATTR);
-        if (expectedCreatedTime != null && !expectedCreatedTime.equals(eventBusCreatedTime(bus))) {
-            throw new AwsException("ValidationError",
-                    "EventBus ownership changed; refusing to delete: " + busName, 400);
-        }
-        deleteEventBusSafe(busName, region);
-    }
-
-
-    private void provisionEventBusPolicy(StackResource r, JsonNode props, CloudFormationTemplateEngine engine,
-                                         String region) {
-        String busName = resolveOrDefault(props, "EventBusName", engine, "default");
-        String statementId = resolveOptional(props, "StatementId", engine);
-        if (statementId == null || statementId.isBlank()) {
-            throw new AwsException("ValidationException", "EventBusPolicy StatementId is required.", 400);
-        }
-
-        if (props != null && props.has("Statement") && props.get("Statement").isObject()) {
-            // Statement form: merge the full statement into the bus policy, keyed by Sid,
-            // so multiple EventBusPolicy resources on the same bus coexist.
-            try {
-                ObjectNode statement = (ObjectNode) engine.resolveNode(props.get("Statement")).deepCopy();
-                statement.put("Sid", statementId);
-
-                EventBus bus = eventBridgeService.describeEventBus(busName, region);
-                ObjectNode policy;
-                String current = bus.getPolicy();
-                if (current != null && !current.isBlank()) {
-                    policy = (ObjectNode) objectMapper.readTree(current);
-                } else {
-                    policy = objectMapper.createObjectNode();
-                    policy.put("Version", "2012-10-17");
-                    policy.putArray("Statement");
-                }
-                ArrayNode statements = policy.withArray("Statement");
-                for (int i = 0; i < statements.size(); i++) {
-                    if (statementId.equals(statements.get(i).path("Sid").asText(null))) {
-                        statements.remove(i);
-                        break;
-                    }
-                }
-                statements.add(statement);
-                eventBridgeService.putPermission(busName, null, null, statementId, null,
-                        objectMapper.writeValueAsString(policy), region);
-            } catch (AwsException e) {
-                throw e;
-            } catch (Exception e) {
-                throw new AwsException("ValidationException",
-                        "Invalid EventBusPolicy Statement: " + e.getMessage(), 400);
-            }
-            r.setPhysicalId(busName + "|" + statementId);
-            return;
-        }
-
-        // Individual form: Action + Principal (+ optional Condition {Type, Key, Value}).
-        String action = resolveOptional(props, "Action", engine);
-        String principal = resolveOptional(props, "Principal", engine);
-        String conditionJson = null;
-        if (props != null && props.has("Condition") && !props.get("Condition").isNull()) {
-            JsonNode c = engine.resolveNode(props.get("Condition"));
-            String type = c.path("Type").asText(null);
-            String key = c.path("Key").asText(null);
-            String value = c.path("Value").asText(null);
-            if (type != null && key != null && value != null) {
-                ObjectNode condition = objectMapper.createObjectNode();
-                condition.set(type, objectMapper.createObjectNode().put(key, value));
-                conditionJson = condition.toString();
-            }
-        }
-        eventBridgeService.putPermission(busName, action, principal, statementId, conditionJson, null, region);
-
-        r.setPhysicalId(busName + "|" + statementId);
-    }
-
-    private void removeEventBusPolicySafe(String physicalId, String region) {
-        try {
-            int sep = physicalId.lastIndexOf('|');
-            String busName = sep >= 0 ? physicalId.substring(0, sep) : "default";
-            String statementId = sep >= 0 ? physicalId.substring(sep + 1) : physicalId;
-            eventBridgeService.removePermission(busName, statementId, false, region);
-        } catch (Exception e) {
-            LOG.debugv("Could not remove event bus policy {0}: {1}", physicalId, e.getMessage());
-        }
-    }
 
     // ── Batch ────────────────────────────────────────────────────────────────
